refac: 로컬 클래스 적용
